Defer loading of Yii integration until used
Remove un-necessary YiiIntegrationLoader class
Use non-tracing API in Yii integration
Sort integrations in integrations.c
