+++ src/DDTrace/Integrations/IntegrationsLoader.php
@@ -83,8 +83,6 @@ class IntegrationsLoader
             '\DDTrace\Integrations\Symfony\SymfonyIntegration';
         $this->integrations[WordPressIntegration::NAME] =
             '\DDTrace\Integrations\WordPress\WordPressIntegration';
-        $this->integrations[YiiIntegration::NAME] =
-            '\DDTrace\Integrations\Yii\YiiIntegration';
         $this->integrations[ZendFrameworkIntegration::NAME] =
             '\DDTrace\Integrations\ZendFramework\ZendFrameworkIntegration';
 
@@ -98,6 +96,8 @@ class IntegrationsLoader
                 '\DDTrace\Integrations\PDO\PDOIntegration';
             $this->integrations[PredisIntegration::NAME] =
                 '\DDTrace\Integrations\Predis\PredisIntegration';
+            $this->integrations[YiiIntegration::NAME] =
+                '\DDTrace\Integrations\Yii\YiiIntegration';
         }
     }
